api/api_penguins/requirements.txt
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
bcrypt==4.1.2
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
ecdsa==0.19.0
exceptiongroup==1.2.0
fastapi==0.110.1
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
idna==3.7
Jinja2==3.1.3
joblib==1.4.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy @ file:///Users/runner/miniforge3/conda-bld/numpy_1707225416355/work/dist/numpy-1.26.4-cp310-cp310-macosx_11_0_arm64.whl#sha256=14c628943f67e9c10d22faa9c1cdcbf69c3192e88f2f597d03b77a252e589567
packaging==24.0
pandas @ file:///Users/runner/miniforge3/conda-bld/pandas_1712782129530/work
passlib==1.7.4
pillow==10.3.0
protobuf==4.25.3
pyarrow==15.0.2
pyasn1==0.6.0
pydantic==2.7.0
pydantic_core==2.18.1
pydeck==0.8.1b0
Pygments==2.17.2
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1709299778482/work
python-jose==3.3.0
python-multipart==0.0.9
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1706886791323/work
referencing==0.34.0
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
scikit-learn==1.4.2
scipy==1.13.0
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smmap==5.0.1
sniffio==1.3.1
starlette==0.37.2
streamlit==1.33.0
tenacity==8.2.3
threadpoolctl==3.4.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
typing_extensions==4.11.0
tzdata @ file:///home/conda/feedstock_root/build_artifacts/python-tzdata_1707747584337/work
urllib3==2.2.1
uvicorn==0.29.0
